# contracts/agent-api.openapi.yaml
openapi: 3.1.0
info:
  title: Comfy Bridge Agent API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:8787
paths:
  /health:
    get:
      operationId: health
      responses:
        '200':
          description: Process health
  /api/v1/status:
    get:
      operationId: getStatus
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Agent and ComfyUI status }
  /api/v1/workflows:
    get:
      operationId: listWorkflows
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Sanitized workflow catalogue }
  /api/v1/assets:
    post:
      operationId: uploadAsset
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201': { description: Stored asset }
  /api/v1/tasks:
    get:
      operationId: listTasks
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Task snapshots }
    post:
      operationId: createTask
      security: [{ bearerAuth: [] }]
      responses:
        '202': { description: Accepted task }
  /api/v1/tasks/{taskId}:
    get:
      operationId: getTask
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: taskId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Task snapshot }
    delete:
      operationId: deleteTask
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: taskId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Task deleted }
  /api/v1/media/{mediaId}:
    get:
      operationId: getMedia
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: mediaId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Authenticated media stream }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer